pyproject.toml
[tool.poetry]
name = "eyeseg"
version = "0.1.0"
description = ""
authors = ["Olivier Morelle <oli4morelle@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
#wandb = "^0.12.18"
click = "^8.1.3"
eyepie = "^0.6.8"
tensorflow = "^2.11.0"
pandas = "^1.5.2"
tqdm = "^4.64.1"
PyYAML = "^6.0"
tensorflow-addons = "^0.18.0"
tensorflow-probability = "^0.18.0"

[tool.poetry.scripts]
eyeseg = "eyeseg.scripts.main:main"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
pytest = "^7.2.0"
pre-commit = "^2.20.0"
commitizen = "^2.37.0"
yapf = "^0.32.0"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "$version"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
